Added style to Forminput
