+++ src/App.js
@@ -1,11 +1,13 @@
 import React from "react";
 
 import "./App.css";
-import FormInput from "./components/FormInput";
+import FormInput from "./components/Forms/FormInput";
+import Header from "./components/Forms/header";
 
 function App() {
   return (
     <div className="App">
+      <Header />
       <FormInput />
     </div>
   );
