+++ src/components/Forms/header.js
@@ -1,0 +1,15 @@
+import React from "react";
+export default class Header extends React.Component {
+  render() {
+    return (
+      <div>
+        <h3>SANDY-X INC.</h3>
+        <h4>
+          185 Bethridge Road, Etobicoke, Ontario M9W 1N4 Tel: (416) 244-2324
+          Fax: (416) 244-9802 Email: info.sandyx@gmail.com
+        </h4>
+        <h3>APPLICATION FOR PARKING and STORAGE</h3>
+      </div>
+    );
+  }
+}
